+++ docs/source/changelog.rst
@@ -1,6 +1,23 @@
 Changelog
 ---------------
 
+3.0.4 (2016-04-17)
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+- Added functions
+  :func:`__enter__() <pydblite.sqlite.Database.__enter__>` and
+  :func:`__exit__() <pydblite.sqlite.Database.__exit__>` to
+  :class:`Database <pydblite.sqlite.Database>`
+
+3.0.3 (2016-04-02)
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+- Fixed bug where db_filter argument of function
+  :func:`_len() <pydblite.sqlite.Table._len>`,
+  :func:`get_unique_ids() <pydblite.sqlite.Table.get_unique_ids>`, and
+  :func:`get_group_count() <pydblite.sqlite.Table.get_group_count>` was
+  ignored if filter returns 0 matches.
+
 3.0.2 (2015-04-18)
 ~~~~~~~~~~~~~~~~~~~~~~~~